Resolve #187: Fix minimize button not working
This bug was introduced in c874899b90751a9459a5661275894073255aacd7. Specifically, the new UpdateWindowStyle method unconditionally set the WindowState to the saved restoreWindowState, including when the window was minimized. This method was called frequently (including on timer ticks) and would constantly un-minimize the timer window.

This change resolves the issue by not clobbering the minimize WindowState when it is set in UpdateWindowStyle.