Create new "Signed" build configurations
Prior to this change there was an unsigned "Debug" configuration and a signed "Release" configuration (and "Portable" equivalents). This meant that people without access to the code signing certificate could not build release versions of the binaries.

After this change there are unsigned "Debug" and "Release" configurations, as well as a signed "Signed" configuration (which is the same as "Release" except, of course, signed).