Resolve #202: Crash due to Invalid Width property
This work around clamps the `Width` property to the range 0..16 inclusive, which should prevent a crash.

It's not clear to me how this calculation goes wrong in the first place. Most likely things are not running on the threads I think they are and there is some concurrent state modification.